Fixing Issue 1
Added generic root element match to common synthesis script to catch the
root element of any topic that does not have a specific matching element
template of its own. Build now dies with an error if there is no
matching root element rule for a topic type included in the build.