Moved script settings to topic type config files
Needed to introduce a script-style config setting, and attribute for
script element, in order to select the correct toc build for generic
topics. Not really the right mechanism. Reflects the fact that not all
of the presentation layer is topic-type specific. Needs more thought on
how best to handle this.